Move exception handlers to dedicated file
