+++ application/backend/src/main.py
@@ -1,23 +1,16 @@
-from collections import defaultdict
-import http
 import logging
 import os
-from typing import Sequence
 
-from fastapi.exceptions import RequestValidationError
-from fastapi.responses import JSONResponse
-import pydantic
 import uvicorn
-from fastapi import FastAPI, Request, Response, status
-from fastapi.encoders import jsonable_encoder
+from fastapi import FastAPI
 
 from api.camera import router as camera_router
 from api.dataset import router as dataset_router
 from api.hardware import router as hardware_router
 from api.project import router as project_router
 from core import lifespan
+from exception_handlers import register_application_exception_handlers
 from settings import get_settings
-from exceptions import GetiBaseException
 
 logger = logging.getLogger(__name__)
 
@@ -35,118 +28,7 @@ app.include_router(hardware_router)
 app.include_router(camera_router)
 app.include_router(dataset_router)
 
-
-@app.exception_handler(GetiBaseException)
-def handle_base_exception(request: Request, e: GetiBaseException) -> Response:
-    """
-    Base exception handler
-    """
-    response = jsonable_encoder({"error_code": e.error_code, "message": e.message, "http_status": e.http_status})
-    headers: dict[str, str] | None = None
-    # 204 skipped as No Content needs to be revalidated
-    if e.http_status not in [200, 201, 202, 203, 205, 206, 207, 208, 226] and request.method == "GET":
-        headers = {"Cache-Control": "no-cache"}  # always revalidate
-    if e.http_status in [204, 304] or e.http_status < 200:
-        return Response(status_code=int(e.http_status), headers=headers)
-    return JSONResponse(content=response, status_code=int(e.http_status), headers=headers)
-
-
-@app.exception_handler(500)
-async def handle_error(request, exception) -> JSONResponse:  # noqa: ANN001, ARG001
-    """
-    Handler for internal server errors
-    """
-    logger.exception(f"Internal server error: {exception}")
-    headers = {"Cache-Control": "no-cache"}  # always revalidate
-    return JSONResponse(
-        {"internal_server_error": "An internal server error occurred."},
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        headers=headers,
-    )
-
-
-@app.exception_handler(404)
-async def handle_not_found(request, exception) -> JSONResponse:  # noqa: ANN001, ARG001
-    """Handler for 'endpoint not found' errors"""
-    message = f"Endpoint {request.url} is not found"
-    logger.info(message)
-    headers = {"Cache-Control": "no-cache"}  # always revalidate
-    return JSONResponse(
-        {"endpoint_not_found_response": message},
-        status_code=status.HTTP_404_NOT_FOUND,
-        headers=headers,
-    )
-
-
-@app.exception_handler(RequestValidationError)
-async def validation_exception_handler(request: Request, exc: RequestValidationError) -> JSONResponse:  # noqa: ARG001
-    """
-    Converts a RequestValidationError to a better readable Bad request exception.
-    """
-    reformatted_message = defaultdict(list)
-    for pydantic_error in exc.errors():
-        # `loc` usually is a list with 2 items describing the location of the error.
-        # The first item specifies if the error is a body, query or path parameter and
-        # the second is the parameter name. Here, only the parameter name is used along
-        # with a message explaining what the problem with the parameter is.
-        loc, msg = pydantic_error["loc"], pydantic_error["msg"]
-        filtered_loc = loc[1:] if loc[0] in ("body", "query", "path") else loc
-        field_string = ".".join(str(filtered_loc))  # nested fields with dot-notation
-        reformatted_message[field_string].append(msg)
-
-    headers = {"Cache-Control": "no-cache"}  # always revalidate
-    return JSONResponse(
-        status_code=status.HTTP_400_BAD_REQUEST,
-        content=jsonable_encoder(
-            {
-                "error_code": "bad_request",
-                "message": reformatted_message,
-                "http_status": http.HTTPStatus.BAD_REQUEST.value,
-            }
-        ),
-        headers=headers,
-    )
-
-
-@app.exception_handler(pydantic.ValidationError)
-async def pydantic_validation_exception_handler(request: Request, exc: pydantic.ValidationError) -> JSONResponse:  # noqa: ARG001
-    """
-    Converts a pydantic ValidationError to a better readable Bad request exception.
-    """
-
-    def format_location(loc: Sequence[str | int]) -> str:
-        """
-        Format location path with proper dot notation and array indices.
-
-        Example:
-            format_location(['a', 0, 'b', 1, 'c']) -> 'a[0].b[1].c'
-        """
-        result = ""
-        for i, item in enumerate(loc):
-            if isinstance(item, int):
-                result += f"[{item}]"
-            else:
-                result += f".{item}" if i > 0 else str(item)
-        return result
-
-    errors = [
-        {"message": error["msg"], "type": error["type"], "location": format_location(error.get("loc", []))}
-        for error in exc.errors()
-    ]
-
-    headers = {"Cache-Control": "no-cache"}  # always revalidate
-    return JSONResponse(
-        status_code=status.HTTP_400_BAD_REQUEST,
-        content=jsonable_encoder(
-            {
-                "error_code": "invalid_payload",
-                "errors": errors,
-                "http_status": http.HTTPStatus.BAD_REQUEST.value,
-            }
-        ),
-        headers=headers,
-    )
-
+register_application_exception_handlers(app)
 
 if __name__ == "__main__":
     uvicorn_port = int(os.environ.get("HTTP_SERVER_PORT", "7860"))
